In 'small_objects', try to abstract over threading
Whether or not the small-object allocator is thread safe is
controlled by the 'AllocThreading' trait.

This is horrible, because it exposes all the internals
and breaks encapsulation......

I also really don't want this level of code duplication
across the entire crate (which its starting to look like
will happen if we continue down this path)
